Start of painful process of renaming 'tests' to 'exams'
Part 1
